requirements.txt
# BackTestPilot - Google ADK Agent
# Requires Python 3.10+ (3.11+ recommended)

# Core Google AI Platform & GenAI
google-cloud-aiplatform>=1.90.0
google-generativeai>=0.1.0rc1

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0

# Logging
structlog>=23.1.0

# Configuration
python-dotenv>=1.0.0

# API & Deployment
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
